# analytics_sync/tech-doc/openapi.yaml
openapi: 3.1.0
info:
  title: analytics_sync
  version: 0.4.0
  summary: Analytics data sync backend for the Chrome extension
  description: |
    Replaces the retired CloudBase analytics upload path used by the
    `tk-adv-cost-monitor` Chrome extension. Provides two endpoints:

    - `GET /v1/analytics/sync/cursor` — bootstrap the daily job with the
      authoritative `nextRequiredDay` per `(scope, storageKey, campaignId)`.
    - `POST /v1/analytics/sync/batches` — idempotent batch upload with
      per-record outcomes.

    **protocolVersion 2** adds `expectedPageCount` to each record: a day is
    only marked complete (and the cursor only advances) once every page
    `1..expectedPageCount` for that `(sellerId, advertiserId, storageKey,
    campaignId, day)` unit has been durably stored. protocolVersion 1
    requests remain accepted and are treated as implicit single-page days.

    See [tech-doc/architecture.md](architecture.md) for design rationale
    and ambiguity resolutions, and
    [tech-doc/plugin-integration.md](plugin-integration.md) for the
    Chrome extension's wiring instructions.
  contact:
    name: tts-erp maintainers
  license:
    name: internal

servers:
  - url: http://127.0.0.1:9877
    description: local dev (tts_erp_v2 on :9877)
  - url: https://daqiang.nat100.top
    description: production (NAT-public domain, nginx reverse-proxy)

tags:
  - name: cursor
    description: Daily-job bootstrap
  - name: batches
    description: Idempotent batch upload

security:
  - bearerAuth: []

paths:
  /healthz:
    get:
      summary: Liveness probe
      description: Returns 200 with service identity. No auth required.
      security: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [status, service, version]
                properties:
                  status: {type: string, enum: [ok]}
                  service: {type: string}
                  version: {type: string}

  /endpoints:
    get:
      summary: API surface manifest
      description: |
        Lists every endpoint, auth requirement, and the error contract.
        No auth required.
      security: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object

  /v1/analytics/sync/cursor:
    get:
      tags: [cursor]
      summary: Latest-day cursors per (storageKey, campaignId)
      description: |
        Returns the per-`(storageKey, campaignId)` latest persisted day for
        the given scope, plus the authoritative `nextRequiredDay` (which
        the plugin MUST use to drive its daily jobs).

        On first sync (no records), `nextRequiredDay` is the bootstrap
        date: `today_in_shop_tz - ANALYTICS_SYNC_BOOTSTRAP_LOOKBACK_DAYS`
        (default 30).
      parameters:
        - name: sellerId
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 128}
        - name: advertiserId
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 128}
        - name: storageKey
          in: query
          required: false
          schema:
            type: string
            enum: [productAnalyses, sessionAnalyses, campaignChangeLogs]
        - name: campaignId
          in: query
          required: false
          schema: {type: string, maxLength: 128}
        - name: cursor
          in: query
          required: false
          schema: {type: string, maxLength: 4096}
          description: "Opaque pagination cursor. MVP: not used."
        - name: pageSize
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CursorResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "500": {$ref: "#/components/responses/InternalError"}

  /v1/analytics/sync/batches:
    post:
      tags: [batches]
      summary: Idempotent batch upload (protocolVersion 1 and 2)
      description: |
        Inserts up to 100 records (≤ 2 MB body). Each record is either
        `inserted`, `duplicate` (already on file), or `rejected` with a
        permanent error code.

        Atomic: every record in the batch commits or rolls back together
        with the daily-completeness and cursor updates. The cursor only
        advances to a day once **all** pages `1..expectedPageCount` for
        that day are durably stored, and it never skips an incomplete
        earlier day (contiguous-prefix rule, anchored at the first day
        the unit has any record for).

        protocolVersion 2 records must carry `expectedPageCount`; all
        records for the same `(sellerId, advertiserId, storageKey,
        campaignId, day)` unit — within a batch and against previously
        stored batches — must agree on it, or the conflicting records are
        rejected with `PAGE_COUNT_CONFLICT` (retryable=false).
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/BatchRequest"}
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BatchResponse"}
        "400":
          $ref: "#/components/responses/BadRequest"
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "500": {$ref: "#/components/responses/InternalError"}

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ttserp_<role>_<32-char urlsafe> (role is `ro`/`rw`/`admin`)
      description: |
        Sync token issued by `api_keys.py create`. The
        plaintext is shown ONCE at creation. Headers accepted:

        - `Authorization: Bearer <token>`
        - `X-Sync-Token: <token>`

  schemas:
    StorageKey:
      type: string
      enum: [productAnalyses, sessionAnalyses, campaignChangeLogs]
      description: |
        Three-value allowlist (mirrors the Chrome extension's
        `AnalyticsStorageKey` type).

    Scope:
      type: object
      required: [sellerId, advertiserId]
      properties:
        sellerId: {type: string, minLength: 1, maxLength: 128}
        advertiserId: {type: string, minLength: 1, maxLength: 128}
        shopName:
          type: string
          nullable: true
          description: Operator-friendly display name. Optional.

    RecordIn:
      type: object
      required:
        - idempotencyKey
        - storageKey
        - campaignId
        - day
        - page
        - endpoint
        - method
        - response
        - source
        - capturedAt
      properties:
        idempotencyKey:
          type: string
          minLength: 64
          maxLength: 64
          pattern: "^[0-9a-f]{64}$"
          description: |
            sha256 hex of canonical json of
            `{sellerId, advertiserId, storageKey, campaignId, day, page}`.
            Server recomputes and rejects mismatches with SCHEMA_INVALID.
        sourceRecordId:
          type: string
          nullable: true
          description: |
            Local UUID from the plugin. Used for tracing only; NOT a
            cross-retry idempotency key.
        storageKey: {$ref: "#/components/schemas/StorageKey"}
        campaignId: {type: string, minLength: 1, maxLength: 128}
        day:
          type: string
          format: date
          description: Inclusive local calendar date `YYYY-MM-DD` in shop TZ.
        page:
          type: integer
          minimum: 1
          description: |
            Positive integer pagination index. The same
            `(scope, storageKey, campaignId, day)` tuple may carry
            multiple `page` values. For protocolVersion 2, must satisfy
            `1 <= page <= expectedPageCount`.
        expectedPageCount:
          type: integer
          minimum: 1
          nullable: true
          description: |
            **protocolVersion 2: required.** Total number of pages the
            TikTok response spans for this `(scope, storageKey, campaignId,
            day)` unit. All records for the same unit must agree on this
            value — within a batch and against previously stored batches.
            A mismatch is rejected with `PAGE_COUNT_CONFLICT`
            (retryable=false) and never advances the cursor.

            Omitted under protocolVersion 1, where every record is treated
            as an implicit single-page day (`expectedPageCount=1`).

            NOT part of the idempotencyKey canonical form.
        endpoint:
          type: string
          minLength: 1
          maxLength: 512
          description: TikTok API path (e.g. `/oec_ads/...`).
        method: {type: string, minLength: 1, maxLength: 16}
        requestBody:
          type: object
          additionalProperties: true
          nullable: true
        response:
          type: object
          additionalProperties: true
          description: |
            Raw response JSON. Capped at 256 KB per record; over-cap
            records are rejected with RESPONSE_TOO_LARGE.
        source:
          type: string
          minLength: 1
          maxLength: 64
          description: Where the capture originated (e.g. `background_poll`).
        capturedAt:
          type: string
          format: date-time
          description: ISO-8601 UTC with explicit `Z` or `+00:00` suffix.
        schemaVersion:
          type: integer
          minimum: 1
          default: 1

    BatchRequest:
      type: object
      required: [protocolVersion, scope, records]
      properties:
        protocolVersion:
          type: integer
          enum: [1, 2]
          description: |
            Server accepts 1 and 2. Version 2 requires
            `records[].expectedPageCount` and enables page-completeness
            cursor semantics: a day only counts as complete once pages
            `1..expectedPageCount` are all durably stored.
        requestId:
          type: string
          minLength: 1
          maxLength: 128
          description: Client-supplied UUID for tracing. Echoed in response.
        scope: {$ref: "#/components/schemas/Scope"}
        records:
          type: array
          minItems: 1
          maxItems: 100
          items: {$ref: "#/components/schemas/RecordIn"}

    CursorEntry:
      type: object
      required: [storageKey, campaignId, nextRequiredDay]
      properties:
        storageKey: {$ref: "#/components/schemas/StorageKey"}
        campaignId: {type: string}
        latestCompletedDay:
          type: string
          format: date
          nullable: true
          description: |
            Last day of the contiguous run of complete days starting at
            the unit's first recorded day. A day is complete only when
            pages `1..expectedPageCount` are all durably stored. NULL when
            the first known day is not yet complete (or no records exist).
        nextRequiredDay:
          type: string
          format: date
          description: |
            Authoritative. The plugin MUST sync from this day forward.
            Equals `latestCompletedDay + 1` when a complete contiguous
            prefix exists; otherwise the first incomplete day at or after
            the unit's first record; otherwise the bootstrap day
            `today - ANALYTICS_SYNC_BOOTSTRAP_LOOKBACK_DAYS` in shop TZ.

    CursorResponse:
      type: object
      required: [code, requestId, data]
      properties:
        code: {type: integer, enum: [0]}
        requestId: {type: string}
        data:
          type: object
          required: [timezone, items, nextCursor]
          properties:
            timezone:
              type: string
              description: Canonical IANA timezone for the seller.
            items:
              type: array
              items: {$ref: "#/components/schemas/CursorEntry"}
            nextCursor:
              type: [string, "null"]
              description: "Opaque pagination cursor (MVP: null)."

    AcceptedRecord:
      type: object
      required: [idempotencyKey, status]
      properties:
        idempotencyKey: {type: string}
        status:
          type: string
          enum: [inserted, duplicate]
          description: |
            Both are successful. The plugin may mark the local record
            as `synced` on either.

    RejectedRecord:
      type: object
      required: [idempotencyKey, code, message, retryable]
      properties:
        idempotencyKey: {type: string}
        code:
          type: string
          enum:
            - SCHEMA_INVALID
            - RESPONSE_TOO_LARGE
            - PAGE_COUNT_CONFLICT
            - SCOPE_DENIED
        message: {type: string}
        retryable:
          type: boolean
          description: |
            Always `false` for schema/response/scope rejections; the
            client must NOT retry unchanged.

    BatchResponse:
      type: object
      required: [code, requestId, data]
      properties:
        code: {type: integer, enum: [0]}
        requestId: {type: string}
        data:
          type: object
          required: [accepted, rejected]
          properties:
            accepted:
              type: array
              items: {$ref: "#/components/schemas/AcceptedRecord"}
            rejected:
              type: array
              items: {$ref: "#/components/schemas/RejectedRecord"}

    ErrorEnvelope:
      type: object
      required: [code, message, requestId, retryable]
      properties:
        code:
          type: string
          description: |
            - `MALFORMED_JSON`
            - `SCHEMA_INVALID`
            - `UNSUPPORTED_PROTOCOL_VERSION`
            - `SCOPE_DENIED`
            - `PAYLOAD_TOO_LARGE`
            - `RATE_LIMITED`
            - `INTERNAL_ERROR`
        message: {type: string}
        requestId: {type: string}
        retryable:
          type: boolean
          description: |
            `true` for transient errors (rate limit, server error);
            `false` for client-correctable errors (4xx).

  responses:
    BadRequest:
      description: Malformed JSON or invalid schema
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}

    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}

    Forbidden:
      description: Token does not grant access to this scope
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}

    PayloadTooLarge:
      description: Body exceeds 2 MB
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}

    RateLimited:
      description: Per-token rate limit exceeded
      headers:
        Retry-After:
          schema: {type: integer}
          description: Seconds until next request is allowed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}

    InternalError:
      description: Server-side failure
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}